fix: resolve comments
